chore: move samples to code-examples dir
