+++ code-examples/sdk/typescript/src/selfservice/logout/express.ts
