+++ code-examples/sdk/typescript/src/selfservice/logout/native.ts
