+++ docs/kratos/self-service/flows/code/logout/index.js
@@ -1,18 +1,18 @@
 // Copyright © 2022 Ory Corp
 // SPDX-License-Identifier: Apache-2.0
 
-import goAPIExample from "!!raw-loader!@site/code-examples/sdk/go/selfservice/logout.go"
-import tsAPIExample from "!!raw-loader!@site/code-examples/sdk/typescript/src/selfservice/logout.ts"
+import golang from "!!raw-loader!@site/code-examples/sdk/go/selfservice/logout.go"
+import native from "!!raw-loader!@site/code-examples/sdk/typescript/src/selfservice/logout/native.ts"
 
 export const apiFlow = {
   ts: {
     label: "TypeScript",
     language: "ts",
-    code: tsAPIExample,
+    code: native,
   },
   go: {
     label: "Go",
     language: "go",
-    code: goAPIExample,
+    code: golang,
   },
 }
