+++ docs/kratos/self-service/flows/user-logout.mdx
@@ -8,12 +8,10 @@ import { apiFlow } from "./code/logout"
 import CodeTabs from "@theme/Code/CodeTabs"
 import Tabs from "@theme/Tabs"
 import TabItem from "@theme/TabItem"
-import browserExample from "!!raw-loader!./code/logout/samples/browser/node.ts"
-import spaExample from "!!raw-loader!./code/logout/samples/spa/react.txt"
 import CodeBlock from "@theme/CodeBlock"
 ```
 
-The logout flow is used to terminate an Ory Identities session. Ory supports logout on all supported clients:
+The logout flow is used to terminate an Ory Identities session. Ory provides logout on all supported clients:
 
 - [Server-rendered browser applications](#server-rendered-browser-applications)
 - [Single page application (SPA)](#single-page-application-spa)
@@ -29,7 +27,9 @@ they want to log out and cache the logout URL. You can then use the cached URL w
 out, for example, accesses a designated UI element.
 
 ```mdx-code-block
-<CodeBlock language="ts" title="index.ts">{browserExample}</CodeBlock>
+import express from "!!raw-loader!@site/code-examples/sdk/typescript/src/selfservice/logout/express.ts"
+
+<CodeBlock language="ts" title="express-logout.ts">{express}</CodeBlock>
 ```
 
 ### Post-logout redirect
@@ -39,42 +39,40 @@ logout request URL. If the parameter is omitted, the default post logout return 
 
 To set the default post-logout return URL:
 
-```mdx-code-block
+````mdx-code-block
 <Tabs group="console-or-cli">
-  <TabItem value="console" label="Ory Console" default>
-```
+<TabItem value="console" label="Ory Console" default>
 
 Go to **Browser Redirects** and set the URL in the **Post-Logout Redirect** field in the
 [Ory Console](https://console.ory.sh/projects/current/browser-redirects).
 
-```mdx-code-block
-  </TabItem>
-  <TabItem value="cli" label="Ory CLI">
-```
+</TabItem>
+<TabItem value="cli" label="Ory CLI">
 
 Run this command in the [Ory CLI](../../../guides/cli/01_installation.mdx):
 
 ```shell
 ory patch identity-config {project-id} \
-   --replace '/selfservice/flows/logout/after/default_browser_return_url="{return_url}"'
+   --replace '/selfservice/flows/logout/after/default_browser_return_url="{return-url}"'
 ```
 
-```mdx-code-block
-  </TabItem>
+</TabItem>
 </Tabs>
-```
+````
 
 If an error occurs, the browser is redirected to the [Error UI](./user-facing-errors.md).
 
 ## Single page application (SPA)
 
-Similar to server-side browser applications, Ory Identities first generates a logout URL for a given Ory Session Cookie. However,
+Similar to server-side browser applications, Ory Identities first generates a logout URL for a given Ory Session cookie. However,
 you can simply call the logout URL using an AJAX request.
 
 The following example requires the [Ory Tunnel](../../../guides/cli/20_proxy-and-tunnel.mdx).
 
 ```mdx-code-block
-<CodeBlock language="ts" title="Logout.tsx">{spaExample}</CodeBlock>
+import react from "!!raw-loader!@site/code-examples/sdk/typescript/src/selfservice/logout/react.tsx"
+
+<CodeBlock language="ts" title="react-logout.tsx">{react}</CodeBlock>
 ```
 
 ## Native API applications
